requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sqlalchemy==2.0.49
aiomysql==0.3.2
pymysql==1.1.3
paho-mqtt==2.1.0
python-dotenv==1.2.2
pydantic==2.13.3
python-jose[cryptography]==3.4.0
pydantic-settings==2.9.1
slowapi==0.1.9
httpx==0.28.1